Refactor audio API into interfaces in order to accomodate XAudio2 backend.

+++ src/synthapp/KeyboardLivePlayInput.cs
@@ -73,7 +73,7 @@ namespace SynthApp
 
         private uint GetCurrentPlaybackSamples()
         {
-            return (uint)_streamSource.SamplesProcessed;
+            return _streamSource.SamplesProcessed;
         }
     }
 }
